moved PostProcess class to postprocessing module
