+++ ai4water/postprocessing/__init__.py
@@ -1,3 +1,4 @@
 
 from .interpret import Interpret
 from .visualize import Visualize
+from .utils import ProcessResults
