+++ ai4water/utils/__init__.py
@@ -1,4 +1,4 @@
 
 from .taylor_diagram import taylor_plot
 from .utils import make_model, prepare_data
-from .visualizations import ProcessResults, murphy_diagram, fdc_plot
+from .visualizations import murphy_diagram, fdc_plot
